+++ docs/conf.py
@@ -67,7 +67,7 @@ release = '1.9.2'
 #
 # This is also used if you do content translation via gettext catalogs.
 # Usually you set "language" from the command line for these cases.
-language = None
+language = 'en'
 
 # Options for localization
 locale_dirs = ['locale/']