+++ src/tiled/propertytypeseditor.h
@@ -20,7 +20,7 @@
 
 #pragma once
 
-#include "properties.h"
+#include "propertytype.h"
 
 #include <QDialog>
 
@@ -29,7 +29,6 @@ class QComboBox;
 class QFormLayout;
 class QItemSelection;
 class QLineEdit;
-class QMenu;
 class QStringListModel;
 class QTreeView;
 
@@ -102,6 +101,7 @@ private:
     void removeValues();
     bool checkValueCount(int count);
 
+    void openClassOfPopup();
     void openAddMemberDialog();
     void addMember(const QString &name, const QVariant &value = QVariant());
     void editMember(const QString &name);
@@ -117,10 +117,17 @@ private:
     void nameEditingFinished();
 
     void colorChanged(const QColor &color);
+    void setDrawFill(bool value);
     void setUsageFlags(int flags, bool value);
     void memberValueChanged(const QStringList &path, const QVariant &value);
 
     void retranslateUi();
+
+    struct NamedFlag {
+        ClassPropertyType::ClassUsageFlag flag;
+        QString name;
+    };
+    QVector<NamedFlag> mFlagsWithNames;
 
     Ui::PropertyTypesEditor *mUi;
     PropertyTypesModel *mPropertyTypesModel;
@@ -134,9 +141,9 @@ private:
 
     ColorButton *mColorButton = nullptr;
     QCheckBox *mUseAsPropertyCheckBox = nullptr;
+    QCheckBox *mDrawFillCheckBox = nullptr;
     QCheckBox *mClassOfCheckBox = nullptr;
     QPushButton *mClassOfButton = nullptr;
-    QMenu *mClassOfMenu;
     QtTreePropertyBrowser *mMembersView = nullptr;
     CustomPropertiesHelper *mPropertiesHelper = nullptr;